move replicate apis into utils
